Change app service name to web
Heroku support said I need to change the name of the service in my Procfile (whatever that is for a Rails app) from app to web, because that's what Heroku is looking for. So I'm going to try changing it in my `docker-compose.yml` file and see if that works.